Refactor Cap.js functions into single handler
